+++ package.json
@@ -20,6 +20,7 @@
     "react-dom": "^16.12.0",
     "react-google-charts": "^3.0.15",
     "react-icons": "^3.8.0",
+    "react-perfect-scrollbar": "^1.5.3",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "react-slick": "^0.25.2",
