+++ src/App.js
@@ -9,14 +9,17 @@ import Routes from './routes';
 import GlobalStyle from './styles/global';
 import theme from './theme';
 
+import 'react-perfect-scrollbar/dist/css/styles.css';
+
 function App() {
   return (
     <BrowserRouter>
       <GlobalStyle />
       <ThemeProvider theme={theme}>
         <CssBaseline />
-        <Main />
-        <Routes />
+        <Main>
+          <Routes />
+        </Main>
       </ThemeProvider>
     </BrowserRouter>
   );
