+++ src/styles/global.js
@@ -8,11 +8,18 @@ export default createGlobalStyle`
     outline: 0;
     box-sizing: border-box;
   }
-  html, body, #root {
+  html {
     min-height: 100%;
+    height: 100%
   }
   body {
     -webkit-font-smoothing: antialiased !important;
+    min-height: 100%;
+    height: 100%
+  }
+  #root {
+    min-height: 100%
+    height: 100%
   }
   body, input, button {
     font: 14px Roboto, sans-serif;
