+++ src/theme/overrides/MuiAppBar.js
@@ -1,6 +1,6 @@
 export default {
   colorPrimary: {
     color: '#fff',
-    backgroundColor: 'rgba(252, 170, 0, 1)',
+    backgroundColor: '#fe7600',
   },
 };
